# docs/openapi.json
{
  "components": {
    "schemas": {
...
      "AnalysisAccepted": {
        "properties": {
          "chunk_count": {
            "title": "Chunk Count",
            "type": "integer"
          },
          "created": {
            "title": "Created",
            "type": "boolean"
          },
          "job": {
            "$ref": "#/components/schemas/JobOut"
          },
          "provider": {
            "additionalProperties": true,
            "title": "Provider",
            "type": "object"
          },
          "truncated": {
            "title": "Truncated",
            "type": "boolean"
          }
        },
        "required": [
          "job",
          "created",
          "truncated",
          "chunk_count",
          "provider"
        ],
        "title": "AnalysisAccepted",
        "type": "object"
      },
      "AnalysisIn": {
        "properties": {
          "scope": {
            "default": "unassigned",
            "enum": [
              "unassigned",
              "all"
            ],
            "title": "Scope",
            "type": "string"
          }
        },
        "title": "AnalysisIn",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "JobOut": {
        "properties": {
          "attempts": {
            "title": "Attempts",
            "type": "integer"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "error_code": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error Code"
          },
          "error_message": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error Message"
          },
          "finished_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Finished At"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "kind": {
            "title": "Kind",
            "type": "string"
          },
          "max_attempts": {
            "title": "Max Attempts",
            "type": "integer"
          },
          "progress": {
            "title": "Progress",
            "type": "integer"
          },
          "result": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Result"
          },
          "started_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Started At"
          },
          "status": {
            "title": "Status",
            "type": "string"
          },
          "status_url": {
            "title": "Status Url",
            "type": "string"
          }
        },
        "required": [
          "id",
          "kind",
          "status",
          "progress",
          "attempts",
          "max_attempts",
          "error_code",
          "error_message",
          "result",
          "created_at",
          "started_at",
          "finished_at",
          "status_url"
        ],
        "title": "JobOut",
        "type": "object"
      },
...
      "Page_JobOut_": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/JobOut"
            },
            "title": "Items",
            "type": "array"
          },
          "limit": {
            "title": "Limit",
            "type": "integer"
          },
          "offset": {
            "title": "Offset",
            "type": "integer"
          },
          "total": {
            "title": "Total",
            "type": "integer"
          }
        },
        "required": [
          "items",
          "total",
          "limit",
          "offset"
        ],
        "title": "Page[JobOut]",
        "type": "object"
      },
...
      "SettingsOut": {
        "properties": {
          "ai_calls_this_month": {
            "title": "Ai Calls This Month",
            "type": "integer"
          },
          "ai_monthly_call_budget": {
            "title": "Ai Monthly Call Budget",
            "type": "integer"
          },
          "allow_self_approval": {
            "title": "Allow Self Approval",
            "type": "boolean"
          },
          "evidence_fresh_days": {
            "title": "Evidence Fresh Days",
            "type": "integer"
          },
          "is_demo": {
            "title": "Is Demo",
            "type": "boolean"
          },
          "max_ai_context_chunks": {
            "title": "Max Ai Context Chunks",
            "type": "integer"
          },
          "max_file_bytes": {
            "title": "Max File Bytes",
            "type": "integer"
          },
          "max_import_rows": {
            "title": "Max Import Rows",
            "type": "integer"
          },
          "max_running_jobs": {
            "title": "Max Running Jobs",
            "type": "integer"
          }
        },
        "required": [
          "max_file_bytes",
          "max_import_rows",
          "max_running_jobs",
          "ai_monthly_call_budget",
          "ai_calls_this_month",
          "max_ai_context_chunks",
          "allow_self_approval",
          "evidence_fresh_days",
          "is_demo"
        ],
        "title": "SettingsOut",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
      "WorkspaceDetail": {
        "properties": {
          "ai": {
            "additionalProperties": true,
            "title": "Ai",
            "type": "object"
          },
          "external_org_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "External Org Id"
          },
          "locale": {
            "title": "Locale",
            "type": "string"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "role": {
            "title": "Role",
            "type": "string"
          },
          "settings": {
            "$ref": "#/components/schemas/SettingsOut"
          },
          "slug": {
            "title": "Slug",
            "type": "string"
          },
          "tenant_id": {
            "format": "uuid",
            "title": "Tenant Id",
            "type": "string"
          },
          "timezone": {
            "title": "Timezone",
            "type": "string"
          }
        },
        "required": [
          "tenant_id",
          "slug",
          "name",
          "locale",
          "timezone",
          "role",
          "settings",
          "ai"
        ],
        "title": "WorkspaceDetail",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "decision-evidence API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/workspaces/{tenant_id}": {
      "get": {
        "operationId": "get_workspace_api_v1_workspaces__tenant_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "tenant_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Tenant Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkspaceDetail"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Workspace",
        "tags": [
          "workspace"
        ]
      }
    },
    "/api/v1/workspaces/{tenant_id}/analysis": {
      "post": {
        "operationId": "start_analysis_api_v1_workspaces__tenant_id__analysis_post",
        "parameters": [
          {
            "in": "path",
            "name": "tenant_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Tenant Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AnalysisIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AnalysisAccepted"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Start Analysis",
        "tags": [
          "jobs",
          "analysis"
        ]
      }
    },
...
    "/api/v1/workspaces/{tenant_id}/jobs": {
      "get": {
        "operationId": "list_jobs_api_v1_workspaces__tenant_id__jobs_get",
        "parameters": [
          {
            "in": "path",
            "name": "tenant_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Tenant Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "status",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "enum": [
                    "queued",
                    "running",
                    "succeeded",
                    "failed"
                  ],
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          },
          {
            "in": "query",
            "name": "kind",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "enum": [
                    "import_commit",
                    "analyze_feedback",
                    "draft_rationale"
                  ],
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Kind"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 25,
              "maximum": 100,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "offset",
            "required": false,
            "schema": {
              "default": 0,
              "maximum": 100000,
              "minimum": 0,
              "title": "Offset",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Page_JobOut_"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Jobs",
        "tags": [
          "jobs",
          "analysis"
        ]
      }
    },
    "/api/v1/workspaces/{tenant_id}/jobs/{job_id}": {
      "get": {
        "operationId": "get_job_api_v1_workspaces__tenant_id__jobs__job_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "job_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Job Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "tenant_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Tenant Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Job",
        "tags": [
          "jobs",
          "analysis"
        ]
      }
    },
    "/api/v1/workspaces/{tenant_id}/jobs/{job_id}/retry": {
      "post": {
        "operationId": "retry_job_api_v1_workspaces__tenant_id__jobs__job_id__retry_post",
        "parameters": [
          {
            "in": "path",
            "name": "job_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Job Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "tenant_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Tenant Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Retry Job",
        "tags": [
          "jobs",
          "analysis"
        ]
      }
    },
...
  }
}